prevnextpage component with props
